requirements_test.txt
celery==5.4.0
coverage==7.5.1
mock>=5.1.0
flake8>=7.0.0
tox>=4.15.0
tox-gh>=1.3.1
django-model-utils>=4.5.1
djangorestframework
model_mommy
